refactor strategy item component for reuse
