+++ src/component/common/StrategyItemContainer/StrategyItemContainer.styles.ts
